+++ rpc/Cargo.toml
@@ -29,10 +29,9 @@ unimock = { version = "0.6.8", optional = true }
 bmp_tracing = { workspace = true }
 clap = { version = "4.6.0", features = ["derive"] }
 wallet = { path = "../wallet" }
-# TODO move this to dev-dependencies:
-# `test-support` is required because `bmp_service.rs` currently uses `Testchain` as a
-# placeholder `ChainApi`. Once `rpc` switches to a real chain backend, drop this feature.
-testenv = { path = "../testenv", features = ["test-support"] }
+# TODO move this to dev-dependencies — `bmp_service.rs` currently uses `Testchain` as a
+# placeholder `ChainApi` until `rpc` switches to a real chain backend.
+testenv = { path = "../testenv" }
 
 [build-dependencies]
 tonic-prost-build = "0.14.5"
