+++ testenv/src/lib.rs
@@ -1,7 +1,7 @@
 //! Bitcoin regtest environment using electrsd with automatic executable downloads
 
 use std::net::SocketAddrV4;
-use std::sync::{Mutex, MutexGuard, PoisonError};
+use std::sync::{Arc, Mutex, MutexGuard, PoisonError};
 use std::time::Duration;
 
 use anyhow::{Context as _, Result};
@@ -13,9 +13,9 @@ use bdk_wallet::bitcoin::address::NetworkChecked;
 use bdk_wallet::bitcoin::key::Secp256k1;
 use bdk_wallet::bitcoin::secp256k1::All;
 use bdk_wallet::bitcoin::{Address, Amount, BlockHash, Network, Transaction, Txid};
+use bdk_wallet::chain::spk_client::{FullScanRequest, FullScanResponse};
 use bmp_tracing::tracing;
-#[cfg(feature = "test-support")]
-use chain::Testchain;
+use chain::{ChainApi, ChainScanner};
 use electrsd::corepc_node::Node;
 use electrsd::electrum_client::{Client, ElectrumApi};
 use electrsd::{ElectrsD, corepc_node};
@@ -27,7 +27,7 @@ use tempfile::TempDir;
 use tokio::net::TcpListener;
 
 /// Bitcoin regtest environment manager
-pub struct TestEnv<'a> {
+pub struct TestEnv {
     bitcoind: Node,
     electrsd: ElectrsD,
     timeout: Duration,
@@ -39,7 +39,7 @@ pub struct TestEnv<'a> {
     explorer_port: Option<u16>,
     bitcoin_rpc_pwd: String,
     mempool: Vec<Txid>,
-    _guard: Option<MutexGuard<'a, ()>>,
+    _guard: Option<MutexGuard<'static, ()>>,
 }
 
 /// Configuration parameters.
@@ -158,7 +158,7 @@ pub fn validate_rpcauth(rpcauth_line: &str, username: &str, password: &str) -> b
 }
 static TESTENV_LOCK: Mutex<()> = Mutex::new(());
 
-impl<'a> TestEnv<'a> {
+impl TestEnv {
     /// Create a new test environment with automatic executable downloads
     pub fn new() -> Result<Self> {
         Self::new_with_conf(Config::default())
@@ -286,7 +286,11 @@ impl<'a> TestEnv<'a> {
     }
 
     /// Create a new [`Testchain`] backed by a fresh electrum client connected to this environment.
-    #[cfg(feature = "test-support")]
+    ///
+    /// Each call returns an independent owned handle (with its own Electrum connection and
+    /// transaction cache), suitable for `Box<dyn ChainApi>` consumers that want one handle per
+    /// simulated peer. For ergonomic single-instance use, `&TestEnv` itself implements
+    /// [`ChainApi`] and [`ChainScanner`].
     pub fn new_testchain(&self) -> Result<Testchain> {
         Ok(Testchain::new(self.new_client()?))
     }
@@ -554,7 +558,81 @@ impl<'a> TestEnv<'a> {
     }
 }
 
-impl<'a> Drop for TestEnv<'a> {
+// Note: `TestEnv` itself cannot implement `ChainApi` because it holds a non-`Send`
+// `MutexGuard` (used to serialize tests by default), but `ChainApi: Send + Sync`. Callers
+// that need an owned `Box<dyn ChainApi>` use [`TestEnv::new_testchain`], which returns a
+// `Testchain` handle with its own electrum client. `ChainScanner`, which has no thread-safety
+// bound, is implemented directly for ergonomic test use (e.g. `wallet.sync_all(&env)`).
+impl ChainScanner for TestEnv {
+    fn populate_tx_cache(&self, txs: impl IntoIterator<Item = impl Into<Arc<Transaction>>>) {
+        self.bdk_electrum_client.populate_tx_cache(txs);
+    }
+
+    fn full_scan<K: Ord + Clone>(
+        &self,
+        request: impl Into<FullScanRequest<K>>,
+        stop_gap: usize,
+        batch_size: usize,
+        fetch_prev_txouts: bool,
+    ) -> anyhow::Result<FullScanResponse<K>> {
+        self.bdk_electrum_client
+            .full_scan(request, stop_gap, batch_size, fetch_prev_txouts)
+            .map_err(Into::into)
+    }
+}
+
+/// Owned Electrum-backed [`ChainApi`] / [`ChainScanner`] handle.
+///
+/// Each instance owns its own [`BdkElectrumClient`] (and therefore its own connection and
+/// transaction cache). Construct via [`TestEnv::new_testchain`] to get a handle connected to
+/// this environment's Electrum endpoint.
+pub struct Testchain {
+    client: BdkElectrumClient<Client>,
+}
+
+impl Testchain {
+    pub const fn new(client: BdkElectrumClient<Client>) -> Self {
+        Self { client }
+    }
+}
+
+impl ChainApi for Testchain {
+    fn transaction_broadcast(&self, tx: &Transaction) -> anyhow::Result<Txid> {
+        broadcast_via(&self.client, tx)
+    }
+}
+
+impl ChainScanner for Testchain {
+    fn populate_tx_cache(&self, txs: impl IntoIterator<Item = impl Into<Arc<Transaction>>>) {
+        self.client.populate_tx_cache(txs);
+    }
+
+    fn full_scan<K: Ord + Clone>(
+        &self,
+        request: impl Into<FullScanRequest<K>>,
+        stop_gap: usize,
+        batch_size: usize,
+        fetch_prev_txouts: bool,
+    ) -> anyhow::Result<FullScanResponse<K>> {
+        self.client
+            .full_scan(request, stop_gap, batch_size, fetch_prev_txouts)
+            .map_err(Into::into)
+    }
+}
+
+/// Shared `ChainApi::transaction_broadcast` body — swallows the idempotent
+/// "Transaction already in block chain" Electrum error and returns the computed txid.
+fn broadcast_via(client: &BdkElectrumClient<Client>, tx: &Transaction) -> anyhow::Result<Txid> {
+    match client.transaction_broadcast(tx) {
+        Ok(txid) => Ok(txid),
+        Err(e) if e.to_string().contains("Transaction already in block chain") => {
+            Ok(tx.compute_txid())
+        }
+        Err(e) => Err(e.into()),
+    }
+}
+
+impl Drop for TestEnv {
     fn drop(&mut self) {
         if let Some(name) = self.container_name.take() {
             tracing::info!("Stopping explorer container {name}...");
