+++ wallet/src/chain_data_source.rs
@@ -1,5 +1,4 @@
 use bdk_wallet::PersistedWallet;
-#[cfg(feature = "test-support")]
 use chain::ChainScanner;
 
 use crate::bmp_wallet::BMPWalletPersister;
@@ -13,29 +12,23 @@ pub trait ChainDataSource {
     -> anyhow::Result<()>;
 }
 
-/// `ChainDataSource` implementation for the Electrum-backed test scaffold [`chain::Testchain`].
+/// `ChainDataSource` implementation for the Electrum-backed [`testenv::Testchain`] handle.
 ///
-/// Gated behind the `test-support` feature so the wallet crate doesn't take a non-test dependency
-/// on the test-only chain backend. The crate's own integration and unit tests activate this
-/// feature via a self-dev-dependency in `Cargo.toml`; downstream crates that exercise the wallet
-/// against a real Electrum node should depend on `wallet` with `features = ["test-support"]`.
-#[cfg(feature = "test-support")]
-impl ChainDataSource for chain::Testchain {
+/// `Testchain` lives in the `testenv` crate alongside `TestEnv`; this impl ties it into the
+/// wallet's sync routine. It mirrors the values previously used by the live Electrum-backed
+/// implementation.
+impl ChainDataSource for testenv::Testchain {
     const RECOVERY_HEIGHT: usize = 190_000;
     const BATCH_SIZE: usize = 16;
-    // Mirror the values previously used by the live Electrum-backed implementation; `Testchain`
-    // is a thin wrapper around a `BdkElectrumClient` used only in integration tests.
     const STOP_GAP: usize = 10;
 
     fn sync(&self, persister: &mut PersistedWallet<impl BMPWalletPersister>) -> anyhow::Result<()> {
-        let stop_gap = Self::STOP_GAP;
-        let batch_size = Self::BATCH_SIZE;
         let tx_nodes = persister.tx_graph().full_txs().map(|tx_node| tx_node.tx);
         self.populate_tx_cache(tx_nodes);
         let request = persister.start_full_scan();
 
         let updates = self
-            .full_scan(request, stop_gap, batch_size, false)
+            .full_scan(request, Self::STOP_GAP, Self::BATCH_SIZE, false)
             .expect("Should be able to start full scan request");
 
         persister.apply_update(updates)?;
